[MWPW-173462] Loading screen: Text should not be marked as a heading (#487)
* Replacing "One moment as we take you to acrobat.adobe.com" h2 tag with
p. Updated css so that the tag change does not affect the text css.

Resolves: [MWPW-173462](https://jira.corp.adobe.com/browse/MWPW-173462)

Testing Notes:
The transition screen UI, specially the font size, style, weight and
line height of the texts should remain same with and without this PR.
We also need to verify the above for backward compatibility. This can be
done for validating the US pages and ROW pages. Authoring changes for US
pages are supposed to be completed within a day or 2, the ROW cahnges
can take some time.
Validation needs to be done for Acrobat as well as Photoshop.

**Test URLs:**
- Before:
https://stage--dc--adobecom.aem.page/acrobat/online/compress-pdf?unitylibs=stage
- After:
https://stage--dc--adobecom.aem.page/acrobat/online/compress-pdf?unitylibs=MWPW-173462