[MWPW-176461] Transition screen progress bar update. (#544)
<!-- Before submitting, please review all open PRs. -->

- This removes the dependency for Milo's `icons.js` to display the
progress bar in Unity transition screens.
- Fixes `progress-bar.svg` 404 issue that was introduced when Milo
updated to federated icons.
- Backwards compatible with icons authoring pattern. Both will work. Re
authoring will need to take place to use the new `[[progress-bar]]`
pattern to transition screens.


Resolves: [MWPW-176461](https://jira.corp.adobe.com/browse/MWPW-176461)

**Test URLs Photoshop:**
- Before:
https://main--cc--adobecom.aem.page/products/photoshop/ai-photo-editor?martech=off
- After:
https://main--cc--adobecom.aem.page/products/photoshop/ai-photo-editor?unitylibs=config-placeholders&martech=off

**Test URLs Acrobat:**
- Before:
https://main--cc--adobecom.aem.page/products/photoshop/ai-photo-editor?martech=off
- After:
https://main--cc--adobecom.aem.page/products/photoshop/ai-photo-editor?unitylibs=config-placeholders&martech=off

--- 

### **Testing notes:**
In order to test the new [[progress-bar]] pattern a separate feature
branch was created that uses updated transition screens that are not
used on production. The test urls listed above use current production
transition screen for Photoshop and Acrobat pages.

To test with the new pattern visit these testing urls. These use an
updated target-config file that we do not want on production, but serve
as testing pages.

-
https://main--cc--adobecom.aem.page/products/photoshop/ai-photo-editor?unitylibs=config-placeholders-poc
-
https://main--dc--adobecom.aem.page/acrobat/online/test/add-pages-to-pdf?unitylibs=config-placeholders-poc

The transition screen should show the progress bar as expected.
No `progress-bar.svg` 404.

Co-authored-by: Ryan Clayton <rclayton@adobe.com>
Co-authored-by: Ruchika Sinha <69535463+Ruchika4@users.noreply.github.com>