Fix for image dimension extraction logic for large images (#635)
Currently we are using image.onload method to decode the image to get
it's dimensions. Image onload method decodes the image in the browser
and thus has memory limitation. For a very large JPG (180 MB), the
browser had to fully decode the image to fire onload. Decoding can
require far more memory than the compressed size, and browsers also
enforce decode/pixel limits. When uploading a large image it was hitting
the browser constraints for memory.

With the new approach, only the first few bytes of the image is decoded
which contains the metadata information.

Testing Notes : 
1. Test with small and large images of each format : jpg, jpeg, png and
webp
2. Regress the LR and Ps workflow testing with images of each format
supported
3. Check the anaytics for image dimension failure

Resolves: [MWPW-182841](https://jira.corp.adobe.com/browse/MWPW-182841)

**Test URLs:**
- Before:
https://main--cc--adobecom.aem.page/drafts/arushi/firefly/firefly?unitylibs=stage
- After:
https://main--cc--adobecom.aem.page/drafts/arushi/firefly/firefly?unitylibs=config-bugs

---------

Co-authored-by: Arushi Gupta <arugupta@Arushis-MacBook-Pro.local>
Co-authored-by: Arushi Gupta <arugupta@arushis-mbp.corp.adobe.com>