[MWPW-182841] : Upload config widget support for firefly (#624)
1. Add support for FF config widget. 
2. Change in Image dimensions extraction logic: Currently we were using
image.onload method to decode the image to get it's dimensions. Image
onload method decodes the image in the browser and thus has memory
limitation. For a very large JPG (180 MB), the browser had to fully
decode the image to fire onload. Decoding can require far more memory
than the compressed size, and browsers also enforce decode/pixel limits.
When uploading a large image it was hitting the browser constraints for
memory. With the new approach, only the first few bytes of the image is
decoded which contains the metadata information.



Resolves: [MWPW-182841](https://jira.corp.adobe.com/browse/MWPW-182841)

**Test URLs:**
- Before:
https://main--cc--adobecom.aem.page/drafts/arushi/firefly/firefly?unitylibs=stage
- After:
https://main--cc--adobecom.aem.page/drafts/arushi/firefly/firefly?unitylibs=ff-upload-widget

Testing Notes:

- File type supported is jpeg/jpg png webp
- Max size supported is 512 MB
- Min file dimensions : 512x512
- No max file dimensions limit

- Test with small and large images of each format : jpg, jpeg, png and
webp
- Regress the LR and Ps workflow testing with images of each format
supported
- Check the anaytics for image dimension failure
- Validate on slow network with large image size
- Cover all major browsers

Authoring Documentation:

https://wiki.corp.adobe.com/display/adobedotcom/CC+Upload+Configurator+for+Firefly

---------

Co-authored-by: Arushi Gupta <arugupta@arushis-mbp.corp.adobe.com>
Co-authored-by: Arushi Gupta <arugupta@Arushis-MacBook-Pro.local>