[MWPW-177099] fix for uploading < uploaded (#650)
fix for uploading < uploaded
1. One such scenario is in case of signed in user for multi file upload.
fixing this.
2. Another guard that I have added before sending the uploaded event is
that we send the the uploaded event only if isUploading is true, in
other words we already sent the uploading event.

Resolves: [MWPW-177099](https://jira.corp.adobe.com/browse/MWPW-177099)

Test URLs:

Before:
https://main--dc--adobecom.aem.live/acrobat/online/compress-pdf?unitylibs=stage
After:
https://main--dc--adobecom.aem.live/acrobat/online/compress-pdf?unitylibs=fixAcrobatAnalytics1

repro steps:
try uploading multiple files for a signed in user on 
https://www.adobe.com/acrobat/online/compress-pdf.html
and see multiple uploaded events being sent.

---------

Co-authored-by: vipulg <vipulg@adobe.com>
Co-authored-by: Arushi Gupta <65466846+arugupta1992@users.noreply.github.com>